requirements.txt
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
Django==2.1.6
djangorestframework==3.9.1
idna==2.8
psycopg2==2.7.7
pycodestyle==2.5.0
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
